+++ testproject/fiber_test/tests/test_all.py
@@ -1,40 +1,28 @@
-import re
 import json
+import time
+from datetime import datetime
+from mock import patch
 
 from django.conf.urls import patterns, url
 from django.views.generic import View
-from django.contrib.auth.models import User
 from django.core.urlresolvers import reverse
 from django.template import Template, Context
-from django.test import TestCase
 from django.core.exceptions import ValidationError
+from django.test import TestCase
+from django.conf import settings
+from django.contrib.auth.models import User
 
-from .models import ContentItem, Page, PageContentItem
-from .utils.validators import FiberURLValidator
-from .mixins import FiberPageMixin
-from .sitemaps import FiberSitemap
+from fiber.utils.validators import FiberURLValidator
+from fiber.mixins import FiberPageMixin
+from fiber.sitemaps import FiberSitemap
+from fiber.models import Page, ContentItem, PageContentItem
+from fiber.utils import date
+from ..test_util import get_short_django_version, format_list, condense_html_whitespace
 
-
-def format_list(l, must_sort=True, separator=' '):
-    """
-    Format a list as a string. Default the items in the list are sorted.
-    E.g.
-    >>> format_list([3, 2, 1])
-    u'1 2 3'
-    """
-    titles = [unicode(v) for v in l]
-    if must_sort:
-        titles = sorted(titles)
-
-    return separator.join(titles)
-
-
-def condense_html_whitespace(s):
-    s = re.sub("\s\s*", " ", s)
-    s = re.sub(">\s*<", "><", s)
-    s = re.sub(" class=\"\s?(.*?)\s?\"", " class=\"\\1\"", s)
-    s = s.strip()
-    return s
+try:
+    from django.utils.timezone import make_aware, utc
+except ImportError:
+    pass
 
 
 class ContentItemTest(TestCase):
@@ -605,7 +593,7 @@ urlpatterns = patterns('',
 
 
 class TestUtilsURLValidator(TestCase):
-    urls = 'fiber.tests'  # use this url conf for our tests
+    urls = urlpatterns  # use this url conf for our tests
     validator = FiberURLValidator()
 
     def test_passes_normal(self):
@@ -733,3 +721,100 @@ class TestSitemap(TestCase):
 
         urls = FiberSitemap().get_urls()
         self.assertEqual(len(urls), 1)
+
+
+
+def mock_tz_now():
+    """
+    Return january 15 2013 10:30
+
+    Depending on the Django version and the settings it will return a datetime with or without timezone.
+    """
+    result = datetime(2013, 1, 15, 10, 30)
+
+    if get_short_django_version() >= (1, 4) and settings.USE_TZ:
+        return make_aware(result, utc)
+    else:
+        return result
+
+
+class FiberTests(TestCase):
+    def test_fiber(self):
+        # setup
+        frontpage = Page.objects.create(title='frontpage', url='/')
+        lorem_ipsum = ContentItem.objects.create(content_html='lorem ipsum')
+        PageContentItem.objects.create(page=frontpage, content_item=lorem_ipsum, block_name='main')
+
+        # - get page
+        response = self.client.get('/')
+        self.assertContains(response, 'lorem ipsum')
+
+    @patch('fiber.utils.date.tz_now', mock_tz_now)
+    def test_friendly_datetime(self):
+        self.assertEqual(date.friendly_datetime(datetime(2013, 1, 15, 10, 30)), 'just now')
+        self.assertEqual(date.friendly_datetime(datetime(2013, 1, 15, 10, 29, 40)), '20 seconds ago')
+        self.assertEqual(date.friendly_datetime(datetime(2013, 1, 15, 10, 29)), 'a minute ago')
+        self.assertEqual(date.friendly_datetime(datetime(2013, 1, 15, 10, 25)), '5 minutes ago')
+        self.assertEqual(date.friendly_datetime(datetime(2013, 1, 15, 9, 30)), 'an hour ago')
+        self.assertEqual(date.friendly_datetime(datetime(2013, 1, 15, 8)), '2 hours ago')
+        self.assertEqual(date.friendly_datetime(datetime(2013, 1, 14, 7)), 'yesterday')
+        self.assertEqual(date.friendly_datetime(datetime(2013, 1, 13)), '2 days ago')
+        self.assertEqual(date.friendly_datetime(datetime(2013, 1, 8)), 'a week ago')
+        self.assertEqual(date.friendly_datetime(datetime(2013, 1, 1)), '2 weeks ago')
+        self.assertEqual(date.friendly_datetime(datetime(2012, 12, 1)), '1 months ago')
+        self.assertEqual(date.friendly_datetime(datetime(2012, 11, 1)), '2 months ago')
+        self.assertEqual(date.friendly_datetime(datetime(2012, 1, 1)), '1 years ago')
+
+        # in the future
+        self.assertEqual(date.friendly_datetime(datetime(2013, 2, 1)), '')
+
+        # invalid input
+        self.assertEqual(date.friendly_datetime('abc'), 'abc')
+
+        # timestamp
+        self.assertEqual(
+            date.friendly_datetime(
+                int(time.mktime(datetime(2013, 1, 15, 10, 29, 20).timetuple()))
+            ),
+            '40 seconds ago'
+        )
+
+    def test_page_view(self):
+        # setup
+        p1 = Page.objects.create(title='p1', url='/p1/')
+        Page.objects.create(title='p2', url='/p2', template_name='template1.html')
+        Page.objects.create(title='p3', url='/p3', is_public=False)
+        Page.objects.create(title='p4', url='/p4', redirect_page=p1)
+
+        # page with default template
+        self.assertContains(self.client.get('/p1/'), '<title>p1</title>')
+
+        # page with custom template
+        self.assertContains(self.client.get('/p2'), 'This is template1.')
+
+        # url without trailing '/'
+        response = self.client.get('/p1')
+        self.assertEqual(response.status_code, 301)
+        self.assertEqual(response['Location'], 'http://testserver/p1/')
+
+        # url does not exist
+        self.assertEqual(self.client.get('/xyz/').status_code, 404)
+
+        # private page
+        self.assertEqual(self.client.get('/p3').status_code, 404)
+
+        # redirect page
+        response = self.client.get('/p4')
+        self.assertEqual(response.status_code, 301)
+        self.assertEqual(response['Location'], 'http://testserver/p1/')
+
+    def test_admin(self):
+        # setup
+        lorem_ipsum = ContentItem.objects.create(content_html='lorem ipsum')
+
+        User.objects.create_superuser('admin', 'admin@ridethepony.nl', 'admin')
+
+        self.client.login(username='admin', password='admin')
+
+        # get admin page for content item
+        self.client.get('/admin/fiber/contentitem/%d/' % lorem_ipsum.id)
